Refactor cached node hook
